requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
python-decouple==3.8
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
boto3==1.34.0
django-storages==1.14.2
requests==2.31.0
openai==1.6.1